Spotless and more apache renaming.
