+++ src/main/java/org/apache/geode/kafka/sink/GeodeKafkaSinkTask.java
@@ -22,13 +22,13 @@ import java.util.stream.Collectors;
 
 import org.apache.kafka.connect.sink.SinkRecord;
 import org.apache.kafka.connect.sink.SinkTask;
-import org.apache.geode.kafka.GeodeContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import org.apache.geode.cache.Region;
 import org.apache.geode.cache.RegionExistsException;
 import org.apache.geode.cache.client.ClientRegionShortcut;
+import org.apache.geode.kafka.GeodeContext;
 
 
 /**
