+++ src/main/java/org/apache/geode/kafka/sink/GeodeSinkConnectorConfig.java
@@ -18,6 +18,7 @@ import java.util.List;
 import java.util.Map;
 
 import org.apache.kafka.common.config.ConfigDef;
+
 import org.apache.geode.kafka.GeodeConnectorConfig;
 
 public class GeodeSinkConnectorConfig extends GeodeConnectorConfig {
