+++ src/main/java/org/apache/geode/kafka/source/GeodeSourceConnectorConfig.java
@@ -19,6 +19,7 @@ import java.util.List;
 import java.util.Map;
 
 import org.apache.kafka.common.config.ConfigDef;
+
 import org.apache.geode.kafka.GeodeConnectorConfig;
 
 public class GeodeSourceConnectorConfig extends GeodeConnectorConfig {
